Fix 'calculateDiscount' method in Motorcycle Class, and fix 'GUIPrincipal'

+++ src/main/java/org/hornetsa/model/IAplicable.java
@@ -1,6 +1,6 @@
 package org.hornetsa.model;
 
 public interface IAplicable {
-    public float calculateDiscount();
+    public double calculateDiscount();
     public void applyDiscont();
 }
